run tests with different stores
